Split gorilla into gorilla, timestamps, and bits
