+++ server/src/types.rs
@@ -41,6 +41,9 @@ pub type Value = std::primitive::f32;
 pub type ArrowValue = datafusion::arrow::datatypes::Float32Type;
 #[cfg(test)] // Proptest is a development dependency.
 pub mod tests {
+    // proptest::num::i64 is not a type and signed integer for compatibility with tables.rs.
+    pub use proptest::num::i64 as ProptestTimestamp;
+
     // proptest::num::f32 is not a type.
     pub use proptest::num::f32 as ProptestValue;
 }
